abstract out createGlobalScope
